switched back to User and Message factory to only cache users for the currently loaded chats
